Finish the review pass over the parallel orchestrator
  Clean up comments.

  Turn the accessor ternaries into early returns, name the stats
  accumulators, and count files as uint64_t from the scan through to
  storage so the clamp and cast at the setter can go.

  Remove the re-entry guard in sftp_parallel_stop, which read a field
  the same function frees, and five stats fields nothing has read since
  development: units_completed_aggregate, endgame_straggler_reaps, and
  the three queue figures. The internal counters that fed only those
  are gone too.

+++ scp.c
@@ -1942,7 +1942,7 @@ toremote(int argc, char **argv, enum scp_mode_e mode, char *sftp_direct)
 					    parallel_want_progress) {
 						char label[64];
 						off_t total_bytes = 0;
-						long total_files = 0, fc;
+						uint64_t total_files = 0, fc;
 						int k;
 
 						for (k = 0; k < argc - 1; k++) {
@@ -1953,8 +1953,9 @@ toremote(int argc, char **argv, enum scp_mode_e mode, char *sftp_direct)
 							total_files += fc;
 						}
 						snprintf(label, sizeof(label),
-						    "Uploading %ld file%s in "
-						    "parallel", total_files,
+						    "Uploading %llu file%s in "
+						    "parallel",
+						    (unsigned long long)total_files,
 						    total_files == 1 ? "" : "s");
 						sftp_parallel_progress_start(
 						    parallel_orch, label,
