+++ sftp.c
@@ -1533,7 +1533,7 @@ process_put(struct sftp_conn *conn, const char *src, const char *dst,
 	if (parallel_orch != NULL && !quiet && g.gl_matchc > 0) {
 		char label[64];
 		off_t total_bytes = 0;
-		long total_files = 0, fc = 0;
+		uint64_t total_files = 0, fc = 0;
 		for (i = 0; g.gl_pathv[i]; i++) {
 			fc = 0;
 			total_bytes += sftp_parallel_scan_upload_total(
@@ -1541,7 +1541,8 @@ process_put(struct sftp_conn *conn, const char *src, const char *dst,
 			total_files += fc;
 		}
 		snprintf(label, sizeof(label),
-		    "Uploading %ld file%s in parallel", total_files,
+		    "Uploading %llu file%s in parallel",
+		    (unsigned long long)total_files,
 		    total_files == 1 ? "" : "s");
 		sftp_parallel_progress_start(parallel_orch, label, total_bytes);
 		sftp_parallel_set_file_total(parallel_orch, total_files);
